requirements.txt
absl-py==1.0.0
anndata==0.7.8
annoy==1.17.0
antlr4-python3-runtime==4.8
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
click==8.0.1
cmake==3.18.4
cycler==0.10.0
Cython==0.29.24
dataclasses==0.6
decorator==4.4.2
distfit==1.4.2
fairseq==0.10.2
fbpca==1.0
filelock==3.0.12
gensim==4.1.2
geosketch==1.2
googledrivedownloader==0.4
h5py==3.5.0
huggingface-hub==0.0.12
hydra-core==1.1.0
idna==3.2
imageio==2.16.0
importlib-metadata==4.6.1
importlib-resources==5.2.0
intervaltree==2.1.0
isodate==0.6.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.37.0
MarkupSafe==2.1.0
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
MulticoreTSNE==0.1
natsort==8.0.0
networkx==2.5.1
nltk==3.6.2
numba==0.54.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264701743/work
olefile==0.46
omegaconf==2.1.0
packaging==21.0
pandas==1.3.0
patsy==0.5.2
pickleshare @ file:///home/hwxu/software/pickleshare-0.7.5-py2.py3-none-any.whl
Pillow==9.0.1
plyfile==0.7.4
portalocker==2.0.0
psutil==5.8.0
pycparser==2.20
pynndescent==0.5.4
pyparsing==2.4.7
pypickle==1.0.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.2.0
PyYAML==5.4.1
rdflib==6.1.1
regex==2021.7.6
requests==2.26.0
rouge-score==0.0.4
sacrebleu==1.5.1
sacremoses==0.0.45
scanorama==1.7.1
scikit-image==0.19.2
scikit-learn==0.24.2
scikit-opt==0.6.5
scipy==1.7.0
seaborn==0.11.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
smart-open==5.2.1
sortedcontainers==2.4.0
statsmodels==0.13.1
threadpoolctl==2.2.0
tifffile==2021.11.2
tokenizers==0.10.3
torch==1.7.1+cu110
torch-cluster @ file:///home/hwxu/torch_cluster-1.5.8-cp37-cp37m-linux_x86_64.whl
torch-geometric==2.0.3
torch-scatter @ file:///home/hwxu/torch_scatter-2.0.5-cp37-cp37m-linux_x86_64.whl
torch-sparse @ file:///home/hwxu/torch_sparse-0.6.8-cp37-cp37m-linux_x86_64.whl
torch-spline-conv @ file:///home/hwxu/torch_spline_conv-1.2.0-cp37-cp37m-linux_x86_64.whl
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.61.2
transformers==4.8.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
umap-learn==0.5.1
urllib3==1.26.6
xlrd==1.2.0
yacs==0.1.8
zipp==3.5.0
